requirements.txt
setuptools
pytest-cov
pytest-watch
ipykernel
pandas
pyyaml
